+++ src/router/index.tsx
@@ -1,10 +1,10 @@
 // @ts-nocheck
 import React from 'react'
 import { createBrowserRouter } from 'react-router-dom'
-import Loading from '../components/Loading'
-const Home = React.lazy(() => import('../index'))
-const Todo = React.lazy(() => import('../pages/Todo'))
-const AntdTable = React.lazy(() => import('../pages/AntdTable'))
+import Loading from '@/components/Loading'
+const Home = React.lazy(() => import('@/index'))
+const Todo = React.lazy(() => import('@/pages/Todo'))
+const AntdTable = React.lazy(() => import('@/pages/AntdTable'))
 const initRouter = [
   {
     path: '/',