+++ webpack.config.js
@@ -1,22 +1,23 @@
 const path = require('path')
+const development = require('./config/webpack.dev.config')
+const production = require('./config/webpack.prod.config')
+const images = require('./config/webpack.image.config')
+const scripts = require('./config/webpack.script.config')
+const styles = require('./config/webpack.styles.config')
 // 生成 htmlWebpackPlugin
 const HtmlWebpackPlugin = require('html-webpack-plugin')
 const webpack = require('webpack')
-const SpeedMeasurePlugin = require('speed-measure-webpack-plugin')
-// 并行压缩
-const TerserPlugin = require('terser-webpack-plugin')
 // fork 线程 check ts type
 const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin')
 // eslint 优化 plugins
 const ESLintPlugin = require('eslint-webpack-plugin')
 // 构建时间
-const smp = new SpeedMeasurePlugin()
+// const SpeedMeasurePlugin = require('speed-measure-webpack-plugin')
+// const smp = new SpeedMeasurePlugin()
 // 压缩 css
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 // 性能分析
 const StatoscopeWebpackPlugin = require('@statoscope/webpack-plugin').default
-// 压缩 html
-const HtmlMinimizerPlugin = require('html-minimizer-webpack-plugin')
 // 压缩
 // const CompressionPlugin = require('compression-webpack-plugin')
 // build 前 清理 dist
@@ -43,7 +44,9 @@ const config = {
 
   plugins: [
     new webpack.HotModuleReplacementPlugin(),
+    // 清除 dist
     new CleanWebpackPlugin(),
+    // htmlWebpackPlugin
     new HtmlWebpackPlugin({
       templateContent: `
 <!DOCTYPE html>
@@ -58,226 +61,26 @@ const config = {
 </html>
     `
     }),
+    // fork ts 检查
     new ForkTsCheckerWebpackPlugin(),
     // eslint plugin
     new ESLintPlugin(),
+    // 压缩 css
     new MiniCssExtractPlugin({ filename: '[name]-[contenthash].css' }),
+    // 计算性能
     new StatoscopeWebpackPlugin()
+    // 压缩
     // new CompressionPlugin()
   ],
   module: {
-    rules: [
-      {
-        test: /\.(gif|png|jpe?g|svg)$/i,
-        // type 属性适用于 Webpack5，旧版本可使用 file-loader
-        exclude: /node_modules/,
-        type: 'asset/resource',
-        use: [
-          {
-            loader: 'image-webpack-loader',
-            options: {
-              // 只在生产环境开启
-              disable: process.env.NODE_ENV === 'development',
-              // jpeg 压缩配置
-              mozjpeg: {
-                quality: 80
-              }
-            }
-          }
-        ]
-      },
-      {
-        test: /\.(png|jpg)$/,
-        exclude: /node_modules/,
-        use: [
-          {
-            loader: 'url-loader',
-            options: {
-              limit: 1024
-            }
-          }
-        ]
-      },
-      {
-        test: /\.svg$/i,
-        use: ['raw-loader']
-      },
-      {
-        test: /\.tsx$/,
-        exclude: /node_modules/,
-        use: [
-          'thread-loader',
-          {
-            loader: 'babel-loader',
-            options: {
-              cacheDirectory: true,
-              presets: [
-                [
-                  '@babel/preset-react',
-                  {
-                    runtime: 'automatic'
-                  }
-                ],
-                '@babel/preset-typescript'
-              ]
-            }
-          }
-        ]
-      },
-      {
-        test: /\.jsx$/,
-        exclude: /node_modules/,
-        use: [
-          'thread-loader',
-          {
-            loader: 'babel-loader',
-            options: {
-              cacheDirectory: true,
-              presets: [
-                [
-                  '@babel/preset-react',
-                  {
-                    runtime: 'automatic'
-                  }
-                ]
-              ]
-            }
-          }
-        ]
-      },
-      {
-        test: /\.less$/,
-        use: [
-          'thread-loader',
-          'style-loader',
-          {
-            loader: 'css-loader',
-            options: {
-              importLoaders: 1
-            }
-          },
-          'less-loader',
-          {
-            loader: 'postcss-loader',
-            options: {
-              postcssOptions: {
-                // 添加 autoprefixer 插件
-                plugins: [require('autoprefixer')]
-              }
-            }
-          }
-        ]
-      },
-      {
-        test: /\.css$/,
-        use: [
-          'thread-loader',
-          'style-loader',
-          {
-            loader: 'css-loader',
-            options: {
-              importLoaders: 1
-            }
-          },
-          {
-            loader: 'postcss-loader',
-            options: {
-              postcssOptions: {
-                // 添加 autoprefixer 插件
-                plugins: [require('autoprefixer')]
-              }
-            }
-          }
-        ]
-      }
-    ]
+    rules: [...images, ...scripts, ...styles]
   }
 }
 
 // 环境配置
 const envConfig = {
-  development: {
-    devtool: 'source-map',
-    // 最小化监听
-    watchOptions: {
-      ignored: /node_modules/
-    },
-    devServer: {
-      hot: true,
-      open: true,
-      historyApiFallback: true // 解决BrowserRouter路由跳转之后刷新浏览器按钮报404的情况
-    },
-    // // 按需编译
-    // experiments: {
-    //   lazyCompilation: true
-    // },
-    // 并行压缩
-    optimization: {
-      removeAvailableModules: false,
-      removeEmptyChunks: false,
-      splitChunks: false,
-      minimize: false,
-      concatenateModules: false,
-      usedExports: false,
-      minimizer: [
-        new TerserPlugin({
-          parallel: 2 // number | boolean
-        })
-      ]
-    }
-  },
-  production: {
-    devtool: 'eval',
-    // cache: {
-    //   type: 'filesystem'
-    // },
-    externals: {
-      lodash: '_',
-      dayjs: 'dayjs'
-    },
-    module: {},
-    module: {},
-    optimization: {
-      splitChunks: {
-        chunks: 'async',
-        minSize: 20000,
-        minRemainingSize: 0,
-        minChunks: 1,
-        maxAsyncRequests: 30,
-        maxInitialRequests: 30,
-        enforceSizeThreshold: 50000,
-        cacheGroups: {
-          defaultVendors: {
-            test: /[\\/]node_modules[\\/]/,
-            priority: -10,
-            reuseExistingChunk: true
-          },
-          default: {
-            minChunks: 2,
-            priority: -20,
-            reuseExistingChunk: true
-          }
-        }
-      },
-      minimize: true,
-      minimizer: [
-        '...',
-        new HtmlMinimizerPlugin({
-          minimizerOptions: {
-            // 折叠 Boolean 型属性
-            collapseBooleanAttributes: true,
-            // 使用精简 `doctype` 定义
-            useShortDoctype: true
-            // ...
-          }
-        }),
-        new TerserPlugin({
-          parallel: 5, // number | boolean
-          exclude: ['node_modules']
-        })
-      ]
-    }
-  }
+  development,
+  production
 }
 
 module.exports = function (e, v) {