+++ config/webpack.prod.config.js
@@ -12,7 +12,6 @@ module.exports = {
     dayjs: 'dayjs'
   },
   module: {},
-  module: {},
   optimization: {
     splitChunks: {
       chunks: 'async',