+++ src/app.tsx
@@ -1,21 +1,9 @@
 import { createRoot } from 'react-dom/client'
 import { router } from './router'
 import { RouterProvider } from 'react-router-dom'
-import { ConfigProvider } from 'antd'
 
-// mobx
-// import { Provider } from 'mobx-react'
-// import store from './store'
-
-import 'antd/dist/antd.variable.min.css' // or 'antd/dist/antd.less'
 import './index.less'
 
 const container = window.document.getElementById('app')!
 const root = createRoot(container) // createRoot(container!) if you use TypeScript
-root.render(
-  <ConfigProvider>
-    {/* <Provider {...store}> */}
-    <RouterProvider router={router} />
-    {/* </Provider> */}
-  </ConfigProvider>
-)
+root.render(<RouterProvider router={router} />)