+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - ✨ 新增(request): 新增 request 请求 ([27f0b42](https://github.com/2401345934/webpack-react-demo/commit/27f0b42))
 - ✨ 新增(router): 改用 history 路由 并修改 webpack 配置 ([5b853bb](https://github.com/2401345934/webpack-react-demo/commit/5b853bb))
 - ✨ 新增(tsconfig): 新增 alias ([e112f53](https://github.com/2401345934/webpack-react-demo/commit/e112f53))
+- ✨ 新增(tsconfig): 修改 tsconfig ([7b2829a](https://github.com/2401345934/webpack-react-demo/commit/7b2829a))
 - ✨ 新增(typescript): 工程化维护 ([3e96241](https://github.com/2401345934/webpack-react-demo/commit/3e96241))
 - ✨ 新增(typescript): 首页 router 纪录 path ([78251c2](https://github.com/2401345934/webpack-react-demo/commit/78251c2))
 - ✨ 新增(typescript): antd router 搭建 ([1c8ef42](https://github.com/2401345934/webpack-react-demo/commit/1c8ef42))